# docs/swagger.yaml
definitions:
  app.Pager:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      totalRows:
        type: integer
    type: object
  errcode.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      msg:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Article:
    properties:
      content:
        type: string
      coverImageURL:
        type: string
      create_by:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      desc:
        type: string
      id:
        type: integer
      state:
        type: integer
      title:
        type: string
      update_by:
        type: string
      updatedAt:
        type: string
    type: object
  model.SwaggerArticle:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
  model.SwaggerTag:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
  model.Tag:
    properties:
      create_by:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      state:
        type: string
      update_by:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 一个完整的博客后端
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/An1ex/BlogService
  title: 博客平台
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: 文章标题
        in: query
        maxLength: 100
        name: title
        type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.SwaggerArticle'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多篇文章
    post:
      parameters:
      - description: 文章标题
        in: body
        maxLength: 100
        minLength: 3
        name: title
        required: true
        schema:
          type: string
      - description: 文章简介
        in: body
        minLength: 3
        name: desc
        schema:
          type: string
      - description: 文章内容
        in: body
        minLength: 3
        name: content
        required: true
        schema:
          type: string
      - description: 文章封面
        in: body
        maxLength: 100
        minLength: 3
        name: cover_image_url
        schema:
          type: string
      - description: 创建者
        in: body
        maxLength: 100
        minLength: 3
        name: created_by
        required: true
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新增文章
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除文章
    get:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文章标题
        in: body
        maxLength: 100
        name: title
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取单篇文章
    put:
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文章标题
        in: body
        maxLength: 100
        minLength: 3
        name: title
        schema:
          type: string
      - description: 文章简介
        in: body
        minLength: 3
        name: desc
        schema:
          type: string
      - description: 文章内容
        in: body
        minLength: 3
        name: content
        schema:
          type: string
      - description: 文章封面
        in: body
        maxLength: 100
        minLength: 3
        name: cover_image_url
        schema:
          type: string
      - description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        required: true
        schema:
          type: integer
      - description: 修改者
        in: body
        maxLength: 100
        minLength: 3
        name: update_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Article'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新文章
  /api/v1/tags:
    get:
      parameters:
      - description: 标签名称
        in: query
        maxLength: 100
        name: name
        type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.SwaggerTag'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个标签
    post:
      parameters:
      - description: 标签名称
        in: body
        maxLength: 100
        minLength: 3
        name: name
        required: true
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      - description: 创建者
        in: body
        maxLength: 100
        minLength: 3
        name: created_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新增标签
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除标签
    get:
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签名称
        in: body
        maxLength: 100
        name: name
        schema:
          type: string
      - default: 1
        description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取单个标签
    put:
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签名称
        in: body
        maxLength: 100
        minLength: 3
        name: name
        schema:
          type: string
      - description: 状态
        enum:
        - 0
        - 1
        in: body
        name: state
        required: true
        schema:
          type: integer
      - description: 修改者
        in: body
        maxLength: 100
        minLength: 3
        name: update_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Tag'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新标签
swagger: "2.0"