# internal/controllers/v1/tag.go
package v1

import (
	"BlogService/global"
	"BlogService/internal/service"
	"BlogService/pkg/app"
	"BlogService/pkg/errcode"

	"github.com/gin-gonic/gin"
	log "github.com/sirupsen/logrus"
)

type Tag struct {
}

func NewTag() Tag {
	return Tag{}
}

// @Summary 获取单个标签
// @Produce  json
// @Param id path int true "标签 ID"
// @Param name body string false "标签名称" maxlength(100)
// @Param state body int false "状态" Enums(0, 1) default(1)
// @Success 200 {object} model.Tag "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags/{id} [get]
func (t Tag) Get(c *gin.Context) {

}

// @Summary 获取多个标签
// @Produce  json
// @Param name query string false "标签名称" maxlength(100)
// @Param state query int false "状态" Enums(0, 1) default(1)
// @Param page query int false "页码"
// @Param page_size query int false "每页数量"
// @Success 200 {object} model.SwaggerTag "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags [get]
func (t Tag) List(c *gin.Context) {
	var param service.TagListRequest
	response := app.NewResponse(c)
	valid, verrs := app.BindAndValid(c, &param)
	if !valid {
		global.Logger.WithFields(log.Fields{
			"errors": verrs,
		}).Error("BindAndValid failed!")
		errRsp := errcode.InvalidParams.WithDetails(verrs.Errors()...)
		response.ToErrorResponse(errRsp)
	} else {
		response.ToResponse(gin.H{})
	}
}

// @Summary 新增标签
// @Produce  json
// @Param name body string true "标签名称" minlength(3) maxlength(100)
// @Param state body int false "状态" Enums(0, 1) default(1)
// @Param created_by body string true "创建者" minlength(3) maxlength(100)
// @Success 200 {object} model.Tag "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags [post]
func (t Tag) Create(c *gin.Context) {

}

// @Summary 更新标签
// @Produce  json
// @Param id path int true "标签 ID"
// @Param name body string false "标签名称" minlength(3) maxlength(100)
// @Param state body int true "状态" Enums(0, 1)
// @Param update_by body string true "修改者" minlength(3) maxlength(100)
// @Success 200 {array} model.Tag "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags/{id} [put]
func (t Tag) Update(c *gin.Context) {

}

// @Summary 删除标签
// @Produce  json
// @Param id path int true "标签 ID"
// @Success 200 {string} string "成功"
// @Failure 400 {object} errcode.Error "请求错误"
// @Failure 500 {object} errcode.Error "内部错误"
// @Router /api/v1/tags/{id} [delete]
func (t Tag) Delete(c *gin.Context) {

}